# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
paths:
  /api/resUsers/{id}:
    get:
      tags:
        - res-user-resource
      operationId: getResUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResUserDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-user-resource
      operationId: updateResUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResUserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-user-resource
      operationId: deleteResUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/resUserRoles/{id}:
    get:
      tags:
        - res-user-role-resource
      operationId: getResUserRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResUserRoleDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-user-role-resource
      operationId: updateResUserRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResUserRoleDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-user-role-resource
      operationId: deleteResUserRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/resUserRoles:
    get:
      tags:
        - res-user-role-resource
      operationId: getAllResUserRoles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResUserRoleDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - res-user-role-resource
      operationId: createResUserRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResUserRoleDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/ApiFieldError"
    ApiFieldError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        property:
          type: string
        rejectedValue:
          type: object
        path:
          type: string
    ResUserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 255
          minLength: 0
        name:
          type: string
          maxLength: 255
          minLength: 0
        role:
          type: array
          items:
            type: integer
            format: int64
      required:
        - username
    ResUserRoleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
...